LayoutTests/imported/w3c/web-platform-tests/css/css-break/table/table-parts-offsetheight-expected.txt


PASS table
FAIL colgroup assert_equals: expected 100 but got 114
FAIL col assert_equals: expected 100 but got 114
FAIL rowgroup assert_equals: expected 100 but got 114
PASS row
PASS cell
PASS content
